chore: rename and update copy for multiple channels with same counterparty alert
