+++ frontend/src/components/channels/ChannelPeerNote.tsx
