+++ frontend/src/screens/channels/IncreaseIncomingCapacity.tsx
@@ -5,7 +5,9 @@ import AppHeader from "src/components/AppHeader";
 import ExternalLink from "src/components/ExternalLink";
 import Loading from "src/components/Loading";
 import { MempoolAlert } from "src/components/MempoolAlert";
-import { PeerAlert } from "src/components/PeerAlert";
+import { ChannelPeerNote } from "src/components/channels/ChannelPeerNote";
+import { ChannelPublicPrivateAlert } from "src/components/channels/ChannelPublicPrivateAlert";
+import { DuplicateChannelAlert } from "src/components/channels/DuplicateChannelAlert";
 import {
   Button,
   ExternalLinkButton,
@@ -32,8 +34,6 @@ import { useChannelPeerSuggestions } from "src/hooks/useChannelPeerSuggestions";
 import { useChannels } from "src/hooks/useChannels";
 import { useInfo } from "src/hooks/useInfo";
 import { cn, formatAmount } from "src/lib/utils";
-import { ChannelPeerNote } from "src/screens/channels/ChannelPeerNote";
-import { ChannelPublicPrivateAlert } from "src/screens/channels/ChannelPublicPrivateAlert";
 import useChannelOrderStore from "src/state/ChannelOrderStore";
 import {
   Channel,
@@ -214,6 +214,7 @@ function NewChannelInternal({
           </div>
         }
       />
+      <MempoolAlert />
       <div className="md:max-w-md max-w-full flex flex-col gap-5 flex-1">
         <img
           src="/images/illustrations/lightning-network-dark.svg"
@@ -408,17 +409,16 @@ function NewChannelInternal({
               Advanced Options
             </Button>
           )}
-          {showAdvanced && (
-            <PeerAlert
-              pubkey={selectedPeer?.pubkey}
-              name={selectedPeer?.name}
-            />
-          )}
-          <MempoolAlert />
           {channels?.some((channel) => channel.public !== !!order.isPublic) && (
             <ChannelPublicPrivateAlert />
           )}
           {selectedPeer?.note && <ChannelPeerNote peer={selectedPeer} />}
+          {showAdvanced && (
+            <DuplicateChannelAlert
+              pubkey={selectedPeer?.pubkey}
+              name={selectedPeer?.name}
+            />
+          )}
           <Button size="lg">Next</Button>
         </form>
