+++ frontend/src/screens/channels/IncreaseOutgoingCapacity.tsx
@@ -5,7 +5,9 @@ import AppHeader from "src/components/AppHeader";
 import ExternalLink from "src/components/ExternalLink";
 import Loading from "src/components/Loading";
 import { MempoolAlert } from "src/components/MempoolAlert";
-import { PeerAlert } from "src/components/PeerAlert";
+import { ChannelPeerNote } from "src/components/channels/ChannelPeerNote";
+import { ChannelPublicPrivateAlert } from "src/components/channels/ChannelPublicPrivateAlert";
+import { DuplicateChannelAlert } from "src/components/channels/DuplicateChannelAlert";
 import {
   Button,
   ExternalLinkButton,
@@ -34,8 +36,6 @@ import { useChannels } from "src/hooks/useChannels";
 import { useInfo } from "src/hooks/useInfo";
 import { usePeers } from "src/hooks/usePeers";
 import { cn, formatAmount } from "src/lib/utils";
-import { ChannelPeerNote } from "src/screens/channels/ChannelPeerNote";
-import { ChannelPublicPrivateAlert } from "src/screens/channels/ChannelPublicPrivateAlert";
 import useChannelOrderStore from "src/state/ChannelOrderStore";
 import {
   Network,
@@ -196,6 +196,7 @@ function NewChannelInternal({ network }: { network: Network }) {
           </div>
         }
       />
+      <MempoolAlert />
       <div className="md:max-w-md max-w-full flex flex-col gap-5 flex-1">
         <img
           src="/images/illustrations/lightning-network-dark.svg"
@@ -371,12 +372,14 @@ function NewChannelInternal({ network }: { network: Network }) {
               </div>
             </div>
           </>
-          <PeerAlert pubkey={order?.pubkey} name={selectedPeer?.name} />
-          <MempoolAlert />
           {channels?.some((channel) => channel.public !== !!order.isPublic) && (
             <ChannelPublicPrivateAlert />
           )}
           {selectedPeer?.note && <ChannelPeerNote peer={selectedPeer} />}
+          <DuplicateChannelAlert
+            pubkey={order?.pubkey}
+            name={selectedPeer?.name}
+          />
           <Button size="lg">{openImmediately ? "Open Channel" : "Next"}</Button>
         </form>
